購入までの実装終了&購入後売切にする処理を追加
