+++ app/views/products/show.html.haml
@@ -23,6 +23,7 @@
             %span 送料込み
           -else
             %span 送料別
+          -# ↓ここに、クレジットカードを登録していない場合に、クレジットカード登録画面に遷移させる条件分岐を追加する
       -if @product.buyer_id == 1
         =link_to "売り切れました", "#", class: "disable-btn"
       -elsif user_signed_in? && @product.buyer_id != 1 && current_user.id != @product.user_id
