--- app/views/purchase/index.html.haml
@@ -1,48 +1,0 @@
-  %h2.buy-content__title
-      購入内容の確認
-    .buy-content__item
-      .buy-content__item__inner
-        .buy-item-main
-          .buy-item-image
-            = image_tag "#{@product.images[0].url}", size: '64x64', class: 'buydetails-contet__image'
-          .buy-item-detail
-            .buy-item-name
-              = @item.name
-              %p.buy-price
-                = "¥#{@product.price.to_s}"
-                %span.shipping-free (税込) 送料込み
-    .buy-content__item
-      %form.buy-form
-        .buy-price-table
-          .buy-price-table__left
-            支払金額
-          .buy-price-table__right
-            = "¥#{@product.price.to_s}"
-    .buy-content__user-info
-      .buy-content__user-info__inner
-        %h3 支払方法
-        .user-info-update
-          = link_to "変更する", "#", calss:"update-btn"
-        .user-info-text
-        - if @default_card_information.blank?
-          %br /
-        - else
-          = "**** **** **** " + "#{@default_card_information.last4}"
-          %br
-          - exp_month = @default_card_information.exp_month.to_s
-          - exp_year = @default_card_information.exp_year.to_s.slice(2,3)
-          = "有効期限 " + exp_month + " / " + exp_year
-          %br
-    .buy-content__user-info
-      .buy-content__user-info__inner              
-        %h3 配送先
-        .user-info-update
-          = link_to "変更する","#", calss:"update-btn"
-        .user-info-text
-          〒111-1111
-          %br
-          大阪府大阪市北区〇〇1-11
-          %br
-          山田太郎
-          = form_tag(action: :pay, method: :post) do
-            %button.buy-button{type:"submit"} 購入する
